requirementx.txt
python-crontab
python-redmine
pyyaml
requests
urllib2